rename .h -> .hpp
